Refactors file locations.
Puts some files in the folders as described in the specification.
