+++ tss-esapi/src/structures/attestation/attest_info.rs
