+++ tss-esapi/src/structures/attestation/certify_info.rs
