+++ tss-esapi/src/structures/attestation/command_audit_info.rs
