+++ tss-esapi/src/structures/attestation/creation_info.rs
