+++ tss-esapi/src/structures/attestation/nv_certify_info.rs
