+++ tss-esapi/src/structures/attestation/nv_digest_certify_info.rs
