+++ tss-esapi/src/structures/attestation/quote_info.rs
