+++ tss-esapi/src/structures/attestation/session_audit_info.rs
