+++ tss-esapi/src/structures/clock/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2022 Contributors to the Parsec project.
+// SPDX-License-Identifier: Apache-2.0
+pub mod clock_info;
+pub mod time_info;
