+++ tss-esapi/src/structures/clock/time_info.rs
