+++ tss-esapi/src/structures/lists/algorithm_property.rs
@@ -29,7 +29,7 @@ impl AlgorithmPropertyList {
     /// Private function that calculates the maximum number
     /// elements allowed in internal storage.
     const fn calculate_max_size() -> usize {
-        crate::structures::capabilitydata::max_cap_size::<TPMS_ALG_PROPERTY>()
+        crate::structures::capability_data::max_cap_size::<TPMS_ALG_PROPERTY>()
     }
 }
