+++ tss-esapi/src/structures/lists/ecc_curves.rs
@@ -46,7 +46,7 @@ impl EccCurveList {
     /// Private function that calculates the maximum number
     /// elements allowed in internal storage.
     const fn calculate_max_size() -> usize {
-        crate::structures::capabilitydata::max_cap_size::<TPM2_ECC_CURVE>()
+        crate::structures::capability_data::max_cap_size::<TPM2_ECC_CURVE>()
     }
 }
