+++ tss-esapi/src/structures/lists/handles.rs
@@ -46,7 +46,7 @@ impl HandleList {
     /// Private function that calculates the maximum number
     /// elements allowed in internal storage.
     const fn calculate_max_size() -> usize {
-        crate::structures::capabilitydata::max_cap_size::<TPM2_HANDLE>()
+        crate::structures::capability_data::max_cap_size::<TPM2_HANDLE>()
     }
 }
