+++ tss-esapi/src/structures/lists/tagged_pcr_property.rs
@@ -44,7 +44,7 @@ impl TaggedPcrPropertyList {
     /// Private function that calculates the maximum number
     /// elements allowed in internal storage.
     const fn calculate_max_size() -> usize {
-        crate::structures::capabilitydata::max_cap_size::<TPMS_TAGGED_PCR_SELECT>()
+        crate::structures::capability_data::max_cap_size::<TPMS_TAGGED_PCR_SELECT>()
     }
 }
