+++ tss-esapi/src/structures/mod.rs
@@ -17,8 +17,8 @@
 /////////////////////////////////////////////////////////
 /// The capabilitydata section
 /////////////////////////////////////////////////////////
-mod capabilitydata;
-pub use self::capabilitydata::CapabilityData;
+mod capability_data;
+pub use self::capability_data::CapabilityData;
 /////////////////////////////////////////////////////////
 /// The names section
 /////////////////////////////////////////////////////////
@@ -176,76 +176,25 @@ pub use ecc::point::EccPoint;
 mod signatures;
 pub use signatures::{EccSignature, RsaSignature};
 /////////////////////////////////////////////////////////
-/// Clock info
+/// Attestation Structures
 /////////////////////////////////////////////////////////
-mod clock_info;
-pub use clock_info::ClockInfo;
+mod attestation;
+pub use attestation::{
+    attest::Attest, attest_info::AttestInfo, certify_info::CertifyInfo,
+    command_audit_info::CommandAuditInfo, creation_info::CreationInfo,
+    nv_certify_info::NvCertifyInfo, nv_digest_certify_info::NvDigestCertifyInfo,
+    quote_info::QuoteInfo, session_audit_info::SessionAuditInfo, time_attest_info::TimeAttestInfo,
+};
 /////////////////////////////////////////////////////////
-/// Certify Info
+/// Clock/Time Structures
 /////////////////////////////////////////////////////////
-mod certify_info;
-pub use certify_info::CertifyInfo;
+mod clock;
+pub use clock::{clock_info::ClockInfo, time_info::TimeInfo};
 /////////////////////////////////////////////////////////
-/// Quote Info
+/// Property Structures
 /////////////////////////////////////////////////////////
-mod quote_info;
-pub use quote_info::QuoteInfo;
-/////////////////////////////////////////////////////////
-/// Time Info
-/////////////////////////////////////////////////////////
-mod time_info;
-pub use time_info::TimeInfo;
-/////////////////////////////////////////////////////////
-/// Time Attest Info
-/////////////////////////////////////////////////////////
-mod time_attest_info;
-pub use time_attest_info::TimeAttestInfo;
-/////////////////////////////////////////////////////////
-/// Command Audit Info
-/////////////////////////////////////////////////////////
-mod command_audit_info;
-pub use command_audit_info::CommandAuditInfo;
-/////////////////////////////////////////////////////////
-/// Session Audit Info
-/////////////////////////////////////////////////////////
-mod session_audit_info;
-pub use session_audit_info::SessionAuditInfo;
-/////////////////////////////////////////////////////////
-/// Creation Info
-/////////////////////////////////////////////////////////
-mod creation_info;
-pub use creation_info::CreationInfo;
-/////////////////////////////////////////////////////////
-/// Nv Certify Info
-/////////////////////////////////////////////////////////
-mod nv_certify_info;
-pub use nv_certify_info::NvCertifyInfo;
-/////////////////////////////////////////////////////////
-/// Nv Digest Certify Info
-/////////////////////////////////////////////////////////
-mod nv_digest_certify_info;
-pub use nv_digest_certify_info::NvDigestCertifyInfo;
-/////////////////////////////////////////////////////////
-/// Attest Info
-/////////////////////////////////////////////////////////
-mod attest_info;
-pub use attest_info::AttestInfo;
-/////////////////////////////////////////////////////////
-/// Attest
-/////////////////////////////////////////////////////////
-mod attest;
-pub use attest::Attest;
-/// Tagged Property
-/////////////////////////////////////////////////////////
-mod tagged_property;
-pub use tagged_property::TaggedProperty;
-/////////////////////////////////////////////////////////
-/// AlgorithmProperty
-/////////////////////////////////////////////////////////
-mod algorithm_property;
-pub use algorithm_property::AlgorithmProperty;
-/////////////////////////////////////////////////////////
-/// TaggedPcrSelect
-/////////////////////////////////////////////////////////
-mod tagged_pcr_select;
-pub use tagged_pcr_select::TaggedPcrSelect;
+mod property;
+pub use property::{
+    algorithm_property::AlgorithmProperty, tagged_pcr_select::TaggedPcrSelect,
+    tagged_property::TaggedProperty,
+};
