+++ tss-esapi/src/structures/property/tagged_pcr_select.rs
